refactor: create renderers for each product type - split root product page into a separate components - move styles to shared file - reduce the amount of unused code in different component types - add useGallery composable - add useProductTabs composable - add canAddToCart method on useCart composable
