+++ packages/theme/modules/catalog/product/components/product-types/grouped/GroupedProductSelector.vue
