+++ packages/theme/modules/catalog/product/components/product-types/styles.scss
@@ -1,0 +1,189 @@
+.product {
+  @include for-desktop {
+    display: flex;
+  }
+
+  &__info {
+    margin: var(--spacer-sm) auto;
+    @include for-desktop {
+      margin: 0;
+      max-width: 32.625rem;
+      padding-left: var(--spacer-lg);
+      width: 100%;
+    }
+  }
+
+  &__header {
+    --heading-title-color: var(--c-link);
+    --heading-title-font-weight: var(--font-weight--bold);
+    --heading-padding: 0;
+    margin: 0 var(--spacer-sm);
+    display: flex;
+    justify-content: space-between;
+    @include for-desktop {
+      --heading-title-font-weight: var(--font-weight--semibold);
+      margin: 0 auto;
+    }
+  }
+
+  &__drag-icon {
+    animation: moveicon 1s ease-in-out infinite;
+    color: var(--c-text-disabled);
+  }
+
+  &__price-and-rating {
+    margin: 0 var(--spacer-sm) var(--spacer-base);
+    align-items: center;
+    @include for-desktop {
+      display: flex;
+      justify-content: space-between;
+      margin: var(--spacer-sm) 0 var(--spacer-lg) 0;
+    }
+  }
+
+  &__rating {
+    display: flex;
+    align-items: center;
+    justify-content: flex-end;
+    margin: var(--spacer-xs) 0 var(--spacer-xs);
+  }
+
+  &__count {
+    @include font(
+        --count-font,
+        var(--font-weight--normal),
+        var(--font-size--sm),
+        1.4,
+        var(--font-family--secondary)
+    );
+    color: var(--c-text);
+    text-decoration: none;
+    margin: 0 0 0 var(--spacer-xs);
+  }
+
+  &__description {
+    @include font(
+        --product-description-font,
+        var(--font-weight--light),
+        var(--font-size--base),
+        1.6,
+        var(--font-family--primary)
+    );
+  }
+
+  &__add-to-cart {
+    margin: var(--spacer-base) var(--spacer-sm) 0;
+    @include for-desktop {
+      margin-top: var(--spacer-2xl);
+    }
+  }
+
+  &__guide,
+  &__compare,
+  &__save {
+    display: block;
+    margin: var(--spacer-xl) 0 var(--spacer-base) auto;
+  }
+
+  &__compare {
+    margin-top: 0;
+  }
+
+  &__tabs {
+    --tabs-title-z-index: 0;
+    margin: var(--spacer-lg) auto var(--spacer-2xl);
+    --tabs-title-font-size: var(--font-size--lg);
+    @include for-desktop {
+      margin-top: var(--spacer-2xl);
+    }
+  }
+
+  &__property {
+    margin: var(--spacer-base) 0;
+
+    &__button {
+      --button-font-size: var(--font-size--base);
+    }
+  }
+
+  &__review {
+    padding-bottom: 24px;
+    border-bottom: var(--c-light) solid 1px;
+    margin-bottom: var(--spacer-base);
+  }
+
+  &__additional-info {
+    color: var(--c-link);
+    @include font(
+        --additional-info-font,
+        var(--font-weight--light),
+        var(--font-size--sm),
+        1.6,
+        var(--font-family--primary)
+    );
+
+    &__title {
+      font-weight: var(--font-weight--normal);
+      font-size: var(--font-size--base);
+      margin: 0 0 var(--spacer-sm);
+
+      &:not(:first-child) {
+        margin-top: 3.5rem;
+      }
+    }
+
+    &__paragraph {
+      margin: 0;
+    }
+  }
+
+  &__additional-actions {
+    display: flex;
+    justify-content: flex-start;
+    margin: 0 var(--spacer-sm);
+    @include for-desktop {
+      margin: 0;
+    }
+  }
+
+  &__gallery {
+    flex: 1;
+  }
+}
+
+.sf-gallery {
+  ::v-deep .glide__slide {
+    flex: none;
+  }
+}
+
+@keyframes moveicon {
+  0% {
+    transform: translate3d(0, 0, 0) rotate(90deg) scale(1, -1);
+  }
+  50% {
+    transform: translate3d(10%, 0, 0) rotate(90deg) scale(1, -1);
+  }
+  100% {
+    transform: translate3d(0, 0, 0) rotate(90deg) scale(1, -1);
+  }
+}
+
+.loading {
+  &--product {
+    padding: var(--spacer-3xl) auto;
+    @include for-desktop {
+      padding-top: 3.75rem;
+      padding-bottom: 3.75rem;
+    }
+  }
+
+  &--product-gallery {
+    padding: var(--spacer-3xl) auto;
+    height: 700px;
+    @include for-desktop {
+      padding-top: 3.75rem;
+      padding-bottom: 3.75rem;
+    }
+  }
+}
