+++ packages/theme/modules/catalog/product/enums/ProductTypeEnum.ts
@@ -1,0 +1,7 @@
+export enum ProductTypeEnum {
+  SIMPLE_PRODUCT = 'SimpleProduct',
+  CONFIGURABLE_PRODUCT = 'ConfigurableProduct',
+  BUNDLE_PRODUCT = 'BundleProduct',
+  GROUPED_PRODUCT = 'GroupedProduct',
+
+}
