--- packages/theme/modules/catalog/product/enums/ProductTypeUnion.ts
@@ -1,1 +1,0 @@
-export type ProductTypeUnion = 'SimpleProduct' | 'ConfigurableProduct' | 'BundleProduct' | 'GroupedProduct';
